% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg.R
\docType{class}
\name{ECG}
\alias{ECG}
\title{ECG class}
\format{An \code{\link{R6Class}} generator object

ibi

Numeric vector for the sequence of interbeat intervals

initialize}
\usage{
ECG
}
\arguments{
\item{data}{numeric vector containing the ECG signal}

\item{samplerate}{numeric indicating the samplerate of the signal in
\code{data}.}

\item{marker}{numeric vector containing the markers for the signal. Must
have the same length as \code{data}.}

\item{name}{character string indicating the name of the new object, e.g.
the subject identifier.}

\item{path}{character string indicating the path where export and import
  functions.

filter

Apply a signal filter to the ECG signal, e.g. signal::butter()}

\item{filt}{an object of class ARMA

export_ecg

Export the ECG signal to a text file with the name "'name'_ecg.txt" to
the path specified in 'path'

import_ibi

Import interbeat intervals from text file}

\item{file}{character string specifing the filename to read. Defaults to
  "'name'_ecg_ibi.txt" if argument is missing

hr

Calculate mean heart rate}

\item{segment}{character string indicating the segment to report the
  heart rate for.

nn50

Calculate the number of successive interbeat intervals that differ by
more than 50 ms}

\item{segment}{character string indicating the segment to report the
  nn50 for.

pnn50

Relative nn50 (as percent of total interbeat intervals)}

\item{segment}{character string indicating the segment to report the
  pnn50 for.

nn20

Calculate the number of successive interbeat intervalss that differ by
more than 20 ms}

\item{segment}{character string indicating the segment to report the
  nn20 for.

pnn20

Relative nn20 (as percent of total interbeat intervals)}

\item{segment}{character string indicating the segment to report the
  pnn20 for.

sdnn

Standard deviation of interbeat intervals}

\item{segment}{character string indicating the segment to report the
  sdnn for.

rmssd

Root mean square of successive differences}

\item{segment}{character string indicating the segment to report the
  rmssd for.

sdsd

Standard deviation of successive differences}

\item{segment}{character string indicating the segment to report the
  sdsd for.

plot_ecg

Plot the ECG signal using dygraphs.}

\item{freq}{numeric indicating the frequency (in Hz) used to recode the
  sample. The lower the value, the higher the plotting perfomance but the
  lower the signal accuracy.

plot_ibi

Plot the interbeat interval sequence using dygraphs.

print

has_ibi

Logical indicating whether interbeat intervals are present or not

get_measure}

\item{name}{character string}

\item{segment}{character string

hr}

\item{data}{numeric vector with sequence of interbeat intervals

subset_ibi}

\item{from}{numeric indicating the start position of the subset in
samples}

\item{to}{numeric indicating the end position of the sumbset in samples}
}
\description{
ECG class
}
\keyword{datasets}

