% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\docType{class}
\name{EDA}
\alias{EDA}
\title{EDA class}
\format{An \code{\link{R6Class}} generator object

initialize}
\usage{
EDA
}
\arguments{
\item{data}{numeric vector containing the ECG signal}

\item{samplerate}{numeric indicating the samplerate of the signal in
\code{data}.}

\item{marker}{numeric vector containing the markers for the signal. Must
have the same length as \code{data}.}

\item{name}{character string indicating the name of the new object, e.g.
the subject identifier.}

\item{path}{character string indicating the path where export and import
functions.}
}
\description{
EDA class
}
\keyword{datasets}

