% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyr6.R
\docType{class}
\name{PHYR6_BASE}
\alias{PHYR6_BASE}
\title{phyr6 base class}
\format{An \code{\link{R6Class}} generator object

data

Numeric vector for the physiological signal

samplerate

Numeric of length one indicating the samplerate of the signal in 'data'

marker
Numeric vector with the same length as 'data' indicating the marker
positions

name

Character string indicating the name of the signal (e.g. subject id)

path

Character string indicating the path where export and import functions
act.

segment_marker

Create a segment for given markers}
\usage{
PHYR6_BASE
}
\arguments{
\item{from}{numeric indicating the start marker for segment(s)}

\item{to}{numeric indicating the end marker for segment(s)}

\item{name}{character string indicating the name for the segment}

\item{check}{logical indicating whether to stop if a marker was not found
  (\code{TRUE}) or create the segment with content \code{NA}
  (\code{FALSE})

segment_samples

Create a segment based on given samples}

\item{from}{numeric indicating the start position for the segment in
samples}

\item{to}{numeric indicating the end position for the segment in samples}

\item{name}{character string indicating the name for the segment
plot_data

segments

List containing the specified segments (of class PHYR6_SEGMENTS)

find_marker

Searches for a given marker in the marker vector}

\item{marker}{numeric marker to search}

\item{marker}{numeric marker to report an error for

find_segment

Searches for a segment with a given name

# TODO: Support whitespace character}

\item{name}{character string indicating the segment to search for}

\item{character}{string. The name of a segment

create_segment

Creates a signle segment}

\item{from}{numeric. The beginning of the segment (in samples)}

\item{to}{numeric. The end of the segment (in samples)}

\item{name}{character string indicating the name of the segment

create_segments

Creates multiple segments}

\item{from}{numeric vector}

\item{to}{numeric vector}

\item{name}{character string indicating the base name of the segments.
  Segments will be named by "name_1", "name_2", ...

resample_data

Returns a resampled copy of the 'data' field. This is e.g. used for
downsampling the signal to increase plotting performance.}

\item{freq}{numeric indicating the new sampling frequency in Hertz}
}
\value{
boolean \code{TRUE} if the marker was found or \code{FALSE} if
  not

error_marker

Throws an error that a marker could not be found

boolean \code{TRUE} if the segment exists or \code{FALSE} if not

error_segment

Throws an error that a segment already exists.
}
\description{
phyr6 base class
}
\keyword{datasets}

