% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{PHYR6_SEGMENT}
\alias{PHYR6_SEGMENT}
\title{Create a segment}
\usage{
PHYR6_SEGMENT(start, end, name)
}
\arguments{
\item{start}{numeric indicating the beginning of the segment (in samples)}

\item{end}{numeric indicating the end of the segment (in samples)}

\item{name}{character string indicating the name of the segment}
}
\value{
An object of class "PHYR6_SEGMENT" with the following elements:
  start:
  end:
  name:
}
\description{
A segment defines a time interval in the signal
}

